Everywhere, Fix: Update package names according to Oracle's conventions
