--- src/main/java/dk/sdu/se_f22/BrandModule/Management/IBim.java
@@ -1,4 +1,0 @@
-package dk.sdu.se_f22.BrandModule.Management;
-
-public interface IBim {
-}
