--- src/main/java/dk/sdu/se_f22/SortingModule/Infrastructure/SortingModule.java
@@ -1,4 +1,0 @@
-package dk.sdu.se_f22.SortingModule.Infrastructure;
-
-public class SortingModule {
-}
