+++ src/main/java/dk/sdu/se_f22/brandmodule/management/Bim.java
@@ -1,0 +1,5 @@
+package dk.sdu.se_f22.brandmodule.management;
+
+public class Bim implements IBim {
+
+}
