+++ src/main/java/dk/sdu/se_f22/brandmodule/management/json/brands.json
