+++ src/main/java/dk/sdu/se_f22/brandmodule/management/services/IIndexingService.java
@@ -1,0 +1,4 @@
+package dk.sdu.se_f22.brandmodule.management.services;
+
+public interface IIndexingService {
+}
