+++ src/main/java/dk/sdu/se_f22/brandmodule/management/services/IJsonService.java
@@ -1,0 +1,9 @@
+package dk.sdu.se_f22.brandmodule.management.services;
+
+import dk.sdu.se_f22.sharedlibrary.models.Brand;
+
+import java.util.List;
+
+public interface IJsonService {
+    public List<Brand> deserializeBrand();
+}
