+++ src/main/java/dk/sdu/se_f22/brandmodule/management/services/IndexingService.java
@@ -1,0 +1,5 @@
+package dk.sdu.se_f22.brandmodule.management.services;
+
+public class IndexingService implements IIndexingService {
+
+}
