+++ src/main/java/dk/sdu/se_f22/sortingmodule/infrastructure/SortingModule.java
@@ -1,0 +1,4 @@
+package dk.sdu.se_f22.sortingmodule.infrastructure;
+
+public class SortingModule {
+}
